+++ modules/storage-page-memory/src/main/java/org/apache/ignite/internal/storage/pagememory/VolatilePageMemoryTableStorage.java
@@ -59,6 +59,7 @@ class VolatilePageMemoryTableStorage extends AbstractPageMemoryTableStorage {
         );
     }
 
+    /** {@inheritDoc} */
     @Override
     public boolean isVolatile() {
         return true;
